requirements.txt

Django==5.0.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Pillow==10.2.0
python-magic==0.4.27
django-storages==1.14.2
drf-yasg==1.21.7
